Revert code to release-1.0
The repository was cloned using the `main` branch, which is not what we
want for the upcoming OCP release. We want to use stable `release-1.0`
branch.

This commit restores all code as it is at the tip of `release-1.0`
(1f045e4), and re-vendors the dependencies. All downstream changes were
kept untouched.